Code redundancy in home page
Home page:
In this, there are separate html script for showing campaigns in desktop
views and mobile view. These views are visible and hidden by using
bootstrap classes.
The problem is, html  for desktop and mobile are perfectly visible and
hidden for desktop and mobile view but in desktop or in mobile view it
loads hidden html also, which is not required.

I solved this with jquery and nearly 650 lines of html are less.